requirements.txt
appdirs==1.4.4
attrs==19.3.0
cached-property==1.5.1
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==1.11.29
django-classy-tags==1.0.0
django-cms==3.6.0
django-filer==1.7.1
django-formtools==2.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
django-sekizai==1.1.0
django-treebeard==4.3.1
djangocms-admin-style==1.2.9
djangocms-attributes-field==1.2.0
djangocms-column==1.10.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-link==2.6.1
djangocms-picture==2.0.8
djangocms-snippet==2.1.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.0.5
djangorestframework==3.11.0
easy-thumbnails==2.7
html5lib==1.1
idna==2.8
iso-639==0.4.5
iso3166==1.0.1
iso8601==0.1.12
isodate==0.6.0
lxml==4.5.1
m3u8==0.6.0
m3u8-dl==0.2.1
m3u8downloader==0.8.6
Pillow==7.1.2
pycryptodome==3.9.8
PySocks==1.7.1
pytube3==9.6.4
pytz==2020.1
requests==2.21.0
requests-toolbelt==0.9.1
six==1.15.0
streamlink==1.4.1
typing-extensions==3.7.4.2
Unidecode==1.1.1
urllib3==1.24.3
webencodings==0.5.1
websocket-client==0.57.0
wells==1.4.1
youtube-dl==2020.6.16.1
zeep==3.4.0
